requirements.txt
ffmpeg==1.4
moviepy==1.0.3
Pillow==9.4.0
requests==2.28.2
